Added the code I wrote for a robotic assignment
This program was written to make my robot navigate a maze and map the maze by always following the left wall. It would also save the location of a 'nest' (dark cell) and return to the maze once the maze was explored.

Any function calls starting with FA_ referenced code written by the university and I did not want to upload it to the public without permission from the authors.